Estimate performance of estimateSurvival function for benchmarking

Estimate performance of estimateSurvival function for benchmarking

Usage

benchmarkCohortSurvival(
  cdm,
  targetSize,
  outcomeSize,
  outcomeDateVariable = "cohort_start_date",
  competingOutcomeSize = NULL,
  competingOutcomeDateVariable = "cohort_start_date",
  censorOnCohortExit = FALSE,
  censorOnDate = NULL,
  followUpDays = Inf,
  strata = NULL,
  eventGap = 30,
  estimateGap = 1,
  minCellCount = 5,
  returnParticipants = FALSE
)

Arguments

cdm

CDM reference

targetSize

number of people in the target cohort table

outcomeSize

number of people in the outcome cohort table

outcomeDateVariable

Variable containing date of outcome event

competingOutcomeSize

number of people in the competing outcome cohort table

competingOutcomeDateVariable

Variable containing date of competing event

censorOnCohortExit

If TRUE, an individual's follow up will be censored at their cohort exit

censorOnDate

if not NULL, an individual's follow up will be censored at the given date

followUpDays

Number of days to follow up individuals (lower bound 1, upper bound Inf)

strata

strata

eventGap

Days between time points for which to report survival estimates. First day will be day zero with risk estimates provided for times up to the end of follow-up, with a gap in days equivalent to eventGap.

estimateGap

vector of time points at which to give survival estimates, if NULL estimates at all times are calculated

minCellCount

The minimum number of events to reported, below which results will be obscured. If 0, all results will be reported.

returnParticipants

Either TRUE or FALSE. If TRUE, references to participants from the analysis will be returned allowing for further analysis.

Value

tibble with performance of estimateSurvival function information, according to the selected input parameters

Examples

# \donttest{
cdm <- mockMGUS2cdm()
#> Warning: ! 4 column in person do not match expected column type:
#> • `person_id` is numeric but expected integer
#> • `gender_concept_id` is numeric but expected integer
#> • `year_of_birth` is numeric but expected integer
#> • `month_of_birth` is numeric but expected integer
#> Warning: ! 1 column in observation_period do not match expected column type:
#> • `person_id` is numeric but expected integer
#> Warning: ! 5 column in visit_occurrence do not match expected column type:
#> • `visit_occurrence_id` is numeric but expected integer
#> • `person_id` is numeric but expected integer
#> • `visit_concept_id` is numeric but expected integer
#> • `visit_start_date` is character but expected date
#> • `visit_end_date` is character but expected date
#> Warning: ! 1 casted column in death_cohort (cohort_set) as do not match expected column
#>   type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 6 casted column in death_cohort (cohort_attrition) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in death_cohort do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 casted column in mgus_diagnosis (cohort_set) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 6 casted column in mgus_diagnosis (cohort_attrition) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in mgus_diagnosis do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 casted column in progression (cohort_set) as do not match expected column
#>   type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 6 casted column in progression (cohort_attrition) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in progression do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 6 casted column in progression_type (cohort_attrition) as do not match
#>   expected column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 2 column in progression_type do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> • `subject_id` is numeric but expected integer
#> Warning: ! 4 column in person do not match expected column type:
#> • `person_id` is numeric but expected integer
#> • `gender_concept_id` is numeric but expected integer
#> • `year_of_birth` is numeric but expected integer
#> • `month_of_birth` is numeric but expected integer
#> Warning: ! 1 column in observation_period do not match expected column type:
#> • `person_id` is numeric but expected integer
#> Warning: ! 1 column in death_cohort do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 column in mgus_diagnosis do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 column in progression do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | ETA:  1s
#> Warning: ! 2 column in progression_type do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> • `subject_id` is numeric but expected integer
cdm$condition_occurrence <- cdm$death_cohort %>%
dplyr::rename("condition_start_date" = "cohort_start_date",
             "condition_end_date" = "cohort_end_date") %>%
             dplyr::compute()
surv_timings <- benchmarkCohortSurvival(
cdm, targetSize = 100, outcomeSize = 20)
#> Warning: ! 4 column in person do not match expected column type:
#> • `person_id` is numeric but expected integer
#> • `gender_concept_id` is numeric but expected integer
#> • `year_of_birth` is numeric but expected integer
#> • `month_of_birth` is numeric but expected integer
#> Warning: ! 1 column in observation_period do not match expected column type:
#> • `person_id` is numeric but expected integer
#> Warning: ! 2 column in benchmark_target do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> • `cohort_definition_id` is numeric but expected integer
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> Warning: ! 2 column in benchmark_outcome do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> • `cohort_definition_id` is numeric but expected integer
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> Warning: ! 4 column in person do not match expected column type:
#> • `person_id` is numeric but expected integer
#> • `gender_concept_id` is numeric but expected integer
#> • `year_of_birth` is numeric but expected integer
#> • `month_of_birth` is numeric but expected integer
#> Warning: ! 1 column in observation_period do not match expected column type:
#> • `person_id` is numeric but expected integer
#> Warning: ! 2 column in benchmark_target do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> • `subject_id` is numeric but expected integer
#> Warning: ! 2 column in benchmark_outcome do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> • `subject_id` is numeric but expected integer
#> Getting overall estimates
# }