Participants contributing to a survival analysis

Participants contributing to a survival analysis

Usage

survivalParticipants(result)

Arguments

result

Result object

Value

References to the study participants contributing to a given analysis

Examples

# \donttest{
cdm <- mockMGUS2cdm()
#> Warning: ! 4 column in person do not match expected column type:
#> • `person_id` is numeric but expected integer
#> • `gender_concept_id` is numeric but expected integer
#> • `year_of_birth` is numeric but expected integer
#> • `month_of_birth` is numeric but expected integer
#> Warning: ! 1 column in observation_period do not match expected column type:
#> • `person_id` is numeric but expected integer
#> Warning: ! 5 column in visit_occurrence do not match expected column type:
#> • `visit_occurrence_id` is numeric but expected integer
#> • `person_id` is numeric but expected integer
#> • `visit_concept_id` is numeric but expected integer
#> • `visit_start_date` is character but expected date
#> • `visit_end_date` is character but expected date
#> Warning: ! 1 casted column in death_cohort (cohort_set) as do not match expected column
#>   type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 6 casted column in death_cohort (cohort_attrition) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in death_cohort do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 casted column in mgus_diagnosis (cohort_set) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 6 casted column in mgus_diagnosis (cohort_attrition) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in mgus_diagnosis do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 casted column in progression (cohort_set) as do not match expected column
#>   type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 6 casted column in progression (cohort_attrition) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in progression do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 6 casted column in progression_type (cohort_attrition) as do not match
#>   expected column type:
#> • `cohort_definition_id` from numeric to integer
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 2 column in progression_type do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> • `subject_id` is numeric but expected integer
#> Warning: ! 4 column in person do not match expected column type:
#> • `person_id` is numeric but expected integer
#> • `gender_concept_id` is numeric but expected integer
#> • `year_of_birth` is numeric but expected integer
#> • `month_of_birth` is numeric but expected integer
#> Warning: ! 1 column in observation_period do not match expected column type:
#> • `person_id` is numeric but expected integer
#> Warning: ! 1 column in death_cohort do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 column in mgus_diagnosis do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 1 column in progression do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> Warning: ! 2 column in progression_type do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> • `subject_id` is numeric but expected integer
surv <- estimateSingleEventSurvival(cdm,
                                    targetCohortTable = "mgus_diagnosis",
                                    outcomeCohortTable = "death_cohort",
                                    returnParticipants = TRUE)
#> Warning: ! 1 column in death_cohort do not match expected column type:
#> • `subject_id` is numeric but expected integer
#> - Getting survival for target cohort 'mgus_diagnosis' and outcome cohort
#> 'death_cohort'
#> Getting overall estimates
#> Warning: ! 5 casted column in survival_participants (cohort_attrition) as do not match
#>   expected column type:
#> • `number_records` from numeric to integer
#> • `number_subjects` from numeric to integer
#> • `reason_id` from numeric to integer
#> • `excluded_records` from numeric to integer
#> • `excluded_subjects` from numeric to integer
#> Warning: ! 1 column in survival_participants do not match expected column type:
#> • `subject_id` is numeric but expected integer
survivalParticipants(surv)
#> # A tibble: 1,384 × 6
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date exposure_id
#>                   <int>      <dbl> <date>            <date>                <int>
#>  1                    1          1 1981-01-01        1981-01-01                1
#>  2                    1          2 1968-01-01        1968-01-01                1
#>  3                    1          3 1980-01-01        1980-01-01                1
#>  4                    1          4 1977-01-01        1977-01-01                1
#>  5                    1          5 1973-01-01        1973-01-01                1
#>  6                    1          6 1990-01-01        1990-01-01                1
#>  7                    1          7 1974-01-01        1974-01-01                1
#>  8                    1          8 1974-01-01        1974-01-01                1
#>  9                    1         10 1981-01-01        1981-01-01                1
#> 10                    1         11 1972-01-01        1972-01-01                1
#> # ℹ 1,374 more rows
#> # ℹ 1 more variable: outcome_id <int>
# }