Add survival information to a cohort table

Add survival information to a cohort table

Usage

addCohortSurvival(
  x,
  cdm,
  outcomeCohortTable,
  outcomeCohortId = 1,
  outcomeDateVariable = "cohort_start_date",
  censorOnCohortExit = FALSE,
  censorOnDate = NULL,
  followUpDays = Inf
)

Arguments

x

cohort table to add survival information

cdm

CDM reference

outcomeCohortTable

The outcome cohort table of interest.

outcomeCohortId

ID of event cohorts to include. Only one outcome (and so one ID) can be considered.

outcomeDateVariable

Variable containing date of outcome event

censorOnCohortExit

If TRUE, an individual's follow up will be censored at their cohort exit

censorOnDate

if not NULL, an individual's follow up will be censored at the given date

followUpDays

Number of days to follow up individuals (lower bound 1, upper bound Inf)

Value

Two additional columns will be added to x. The "time" column will contain number of days to censoring. The "status" column will indicate whether the patient had the event (value: 1), or did not have the event (value: 0)

Examples

cohort <- dplyr::tibble(
cohort_definition_id = c(1,1,1),
subject_id = c(1,2,3),
cohort_start_date = c(as.Date("2020-01-01"),
                      as.Date("2020-01-02"),
                      as.Date("2020-01-01")),
cohort_end_date = c(as.Date("2020-04-01"),
                    as.Date("2020-08-02"),
                    as.Date("2020-03-01"))
)
cdm <- PatientProfiles::mockPatientProfiles(cohort1 = cohort)
cdm$cohort1 <- cdm$cohort1 %>%
  addCohortSurvival(
    cdm = cdm,
    outcomeCohortTable = "cohort2",
    outcomeCohortId = 1
  )